voice-parser/pyproject.toml
[project]
name = "voice-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ai-voice-shared",
    "boto3>=1.40.52",
    "httpx>=0.28.1",
    "openai>=2.3.0",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "twilio>=9.8.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.0",
    "voice-parser",
]

[tool.uv.sources]
voice-parser = { workspace = true }
ai-voice-shared = { path = "../shared-lib", editable = true }

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (isolated, mocked dependencies)",
    "integration: Integration tests (mocked external services, real internal logic)",
    "e2e: End-to-end tests (require deployed infrastructure)",
]
